Only show 10 current users max in a conversation.
If there is more than 10 users in a conversation, then display a link to show the rest of the users.
The rest of the users are presented to the current user as a special kind of message in the flow of the conversation,
so it's not a problem if there is hundreds of users in a conversation because the current user can
scroll the message to see them all.