Fix double notification issue
When a user spawns a convo, the user is technically creating a new one, so the followers will receive
a notification about the newly created convo. Also, the original convo gets a notification about the new spawned convo.
If the original convo is also the personal convo of one follower of the user who spawns the conversation, only one
notification is send, no two.