Remove the link to the user personal convo from the navigation menu
The link is redundant, because there is the "home" link. And is not necessary because only the current user will see it, and the current user must learn that "home == personal_convo".
Also, is a bad idea to allow the display of strings of arbitrary length in the navigation menu (the user name) because they can break the layout.
Also, the user will know that is logged by looking at the small avatar, the edit profile link and the logout link.
We need to solve the problem with users with various accounts with different names but same avatars.