Fix the invites system, with tests.
Before, the invite stuff was pretty broken, you can invite someone to a public convo (for example when you create a convo, all of your followers are automaticly invited), if you make the convo private later, these users can't be able to follow the convo because you can't invite them again, to send them the needed invite with token. Also, now the invite view, will show the users that can really be invited.